Add changes

+++ kylieq-CIS415F19-P2/part3.c
@@ -28,28 +28,28 @@ void handler(int sig_num) {
             CHECK = 1;
             //w = waitpid(PCBS[COUNTER]->pid, &wstatus, WNOHANG);
             while(1) {
-              //  if (w == 0) {
+                if (!wait(NULL)) {
                     if (kill(PCBS[COUNTER]->pid, SIGCONT) == 0) {
                         printf("Process: %d - Received Signal SIGALRM - Continued\n", PCBS[COUNTER]->pid);
                         sleep(1);
                         break;
                     }
-                //}
-                //else {
+                }
+                else {
                     COUNTER++;
-                //}
+                }
             }
             while (1) {
-                //if (w == 0) {
+                if (!wait(NULL)) {
                     if (kill(PCBS[COUNTER]->pid, SIGSTOP) == 0) {
                         printf("Process: %d - Received Signal SIGALRM - Suspended\n", PCBS[COUNTER]->pid);
                         sleep(1);
                         break;
                     }
-                //}
-                //else {
+                }
+                else {
                     COUNTER++;
-                //}
+                }
             }
             //w = waitpid(PCBS[COUNTER]->pid, &wstatus, WNOHANG);
     }
